Filtersystemen gereconstrueerd naar herbruikbare component, bestanden herorderd en schoonmaak van code.
