+++ src/components/Logout/Logout.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../Context/AuthContext';
+import { useAuth } from '../../services/Context/AuthContext';
 import styles from "../shared/searchbar/SearchBar.module.css";
 
 const Logout = () => {
