+++ src/helpers/ProtectedRoute/ProtectedRoute.js
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../Context/AuthContext";
+import { useAuth } from "../../services/Context/AuthContext";
 
 export const ProtectedRoute = ({ children }) => {
     const { user } = useAuth();
