+++ src/pages/Homepage/HomePage.js
@@ -2,7 +2,7 @@ import React from "react";
 import SearchBar from "../../components/shared/searchbar/SearchBar";
 import LoginForm from "../../components/Login/Login";
 import Button from "../../components/shared/button/Button";
-import { useAuth } from "../../components/Context/AuthContext";
+import { useAuth } from "../../services/Context/AuthContext";
 import styles from './Homepage.module.css';
 import {useNavigate} from "react-router-dom";
